app/requirements.txt
flask
websockets
pyserial
